scripts/industry-observer/requirements.txt
feedparser>=6.0
requests>=2.31
beautifulsoup4>=4.12
trafilatura>=1.6
spacy>=3.7
rapidfuzz>=3.5
schedule>=1.2
google-generativeai>=0.3
newspaper4k>=0.9
python-dotenv>=1.0
SQLAlchemy>=2.0
lxml>=5.1
tqdm>=4.66
vaderSentiment>=3.3.2
supabase>=2.0.0
